More fixes and tweaks
